# packages/backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Scroblarr API
  version: 1.0.0
  description: |
    API documentation for Scroblarr - Media scrobbling service for Plex, Jellyfin, and Emby.

    This API allows you to manage users, sync media watching history across services, and configure integrations with TVTime and Trakt.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token (Plex or Jellyfin access token)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Global API key for API authentication
  schemas:
...
    LogEntry:
      type: object
      properties:
        level:
          type: integer
        time:
          type: integer
        msg:
          type: string
        label:
          type: string
          nullable: true
    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        pagination:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
paths:
...
  /api/v1/logs:
    get:
      summary: Get application logs
      tags: [Logs]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warn, error, fatal]
        - name: label
          in: query
          schema:
            type: string
            enum: [webhook, sync, auth, api, database, tvtime, plex, system, migration]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/logs/files:
    get:
      summary: Get list of log files
      tags: [Logs]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of log files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        size:
                          type: integer
                        modified:
                          type: string
                          format: date-time
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/logs/download/{filename}:
    get:
      summary: Download log file
      tags: [Logs]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Log file
          content:
            text/plain: {}
            application/gzip: {}
        '400':
          description: Invalid filename
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/avatars/jellyfin/{userId}:
    get:
      summary: Get Jellyfin user avatar
      tags: [Avatars]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avatar image
          content:
            image/jpeg: {}
            image/png: {}
        '400':
          description: Invalid user ID
        '503':
          description: Jellyfin not configured
        '500':
          description: Internal server error
  /api/v1/avatars/plex/{userId}:
    get:
      summary: Get Plex user avatar
      tags: [Avatars]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avatar image
          content:
            image/jpeg: {}
            image/png: {}
        '302':
          description: Redirect to public Plex CDN URL
        '403':
          description: Plex authentication required
        '404':
          description: User or avatar not found
        '500':
          description: Internal server error
  /api/v1/avatars/trakt/{userId}:
    get:
      summary: Get Trakt user avatar
      tags: [Avatars]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avatar image
          content:
            image/jpeg: {}
            image/png: {}
        '404':
          description: User or avatar not found
        '500':
          description: Internal server error
  /api/v1/avatars/tvtime/{userId}:
    get:
      summary: Get TVTime user avatar
      tags: [Avatars]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Avatar image
          content:
            image/jpeg: {}
            image/png: {}
        '404':
          description: User or avatar not found
        '500':
          description: Internal server error